Inline mocks into form.test.js
